feat(donation): rename <DonationArea> to <SupportWidget>
