--- src/views/ArticleDetail/DonationArea/DonationButton/styles.css
@@ -1,12 +1,0 @@
-:root {
-  --donation-area-button-size: 2.5rem;
-}
-
-.container {
-  @mixin flex-center-all;
-
-  position: absolute;
-  top: calc(var(--donation-area-button-size) / 2 * -1);
-  right: 0;
-  left: 0;
-}
