--- src/views/ArticleDetail/DonationArea/styles.css
@@ -1,28 +1,0 @@
-:root {
-  --donation-area-button-size: 2.5rem;
-  --donation-area-avatar-size: 1.25rem;
-}
-
-.container {
-  position: relative;
-  padding: calc(var(--spacing-base) + var(--donation-area-button-size) / 2)
-    var(--spacing-base)
-    calc(var(--spacing-base) + var(--donation-area-avatar-size) / 2);
-  margin: calc(
-      var(--spacing-x-loose) + var(--spacing-base) +
-        var(--donation-area-button-size) / 2
-    )
-    0
-    calc(
-      var(--spacing-x-loose) + var(--spacing-base) +
-        var(--donation-area-avatar-size) / 2
-    );
-  text-align: center;
-  background: var(--color-yellow-lighter);
-  border-radius: var(--spacing-x-tight);
-}
-
-p {
-  font-size: var(--font-size-xs);
-  line-height: 1.25rem;
-}
