+++ src/views/ArticleDetail/SupportWidget/styles.css
@@ -1,0 +1,39 @@
+:root {
+  --donation-area-button-height: 2.5rem;
+  --donation-area-avatar-height: 1.25rem;
+}
+
+.support-widget {
+  position: relative;
+  padding: calc(var(--spacing-base) + var(--donation-area-button-height) / 2)
+    var(--spacing-base)
+    calc(var(--spacing-base) + var(--donation-area-avatar-height) / 2);
+  margin: calc(
+      var(--spacing-x-loose) + var(--spacing-base) +
+        var(--donation-area-button-height) / 2
+    )
+    0 calc(var(--spacing-loose) + var(--donation-area-avatar-height) / 2);
+  text-align: center;
+  background: var(--color-yellow-lighter);
+  border-radius: var(--spacing-x-tight);
+}
+
+p {
+  font-size: var(--font-size-xs);
+  font-weight: var(--font-weight-light);
+  line-height: 1.25rem;
+}
+
+.donation-button {
+  position: absolute;
+  top: calc(var(--donation-area-button-height) / 2 * -1);
+  right: 0;
+  left: 0;
+}
+
+.donators {
+  position: absolute;
+  right: 0;
+  bottom: calc(var(--donation-area-avatar-height) / 2 * -1);
+  left: 0;
+}
